# manager-zedaapi/public/openapi-content-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zé da API - Content Management API",
    "description": "API completa para gerenciamento externo de conteudo da plataforma Zé da API.\n\nTodos os endpoints requerem autenticacao via header `Authorization: Bearer <CONTENT_API_KEY>`.\n\nA chave e definida na variavel de ambiente `CONTENT_API_KEY` do servidor.",
    "version": "1.0.0",
    "contact": {
      "name": "Zé da API",
      "url": "https://zedaapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://app.zedaapi.com",
      "description": "Producao"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CONTENT_API_KEY definida no .env do servidor. Minimo 32 caracteres."
      }
    },
    "schemas": {
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5abc123def456"
          },
          "title": {
            "type": "string",
            "example": "Como integrar WhatsApp API em 5 minutos"
          },
          "slug": {
            "type": "string",
            "example": "como-integrar-whatsapp-api-em-5-minutos"
          },
          "content": {
            "type": "string",
            "example": "<h2>Introducao</h2><p>Neste artigo voce vai aprender...</p>"
          },
          "excerpt": {
            "type": "string",
            "nullable": true,
            "example": "Aprenda a integrar a WhatsApp API rapidamente com a Zé da API."
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://s3.example.com/blog/2026/02/cover.webp"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "example": "draft"
          },
          "seoTitle": {
            "type": "string",
            "nullable": true,
            "example": "Integrar WhatsApp API - Guia Completo 2026"
          },
          "seoDescription": {
            "type": "string",
            "nullable": true,
            "example": "Guia passo a passo para integrar a WhatsApp Business API usando a Zé da API."
          },
          "seoKeywords": {
            "type": "string",
            "nullable": true,
            "example": "whatsapp api, integracao, tutorial, webhook"
          },
          "readingTimeMin": {
            "type": "integer",
            "example": 5
          },
          "viewCount": {
            "type": "integer",
            "example": 142
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/BlogCategory"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "example": "Admin"
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "tagId": {
                  "type": "string"
                },
                "postId": {
                  "type": "string"
                },
                "tag": {
                  "$ref": "#/components/schemas/BlogTag"
                }
              }
            }
          }
        }
      },
      "BlogCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5cat001"
          },
          "name": {
            "type": "string",
            "example": "Tutoriais"
          },
          "slug": {
            "type": "string",
            "example": "tutoriais"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Guias passo a passo para usar a plataforma"
          },
          "sortOrder": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "integer",
                "example": 12
              }
            }
          }
        }
      },
      "BlogTag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5tag001"
          },
          "name": {
            "type": "string",
            "example": "WhatsApp"
          },
          "slug": {
            "type": "string",
            "example": "whatsapp"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "integer",
                "example": 8
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "API key ausente ou invalida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "summary": "Header ausente",
                "value": {
                  "error": "Missing Authorization header"
                }
              },
              "invalid": {
                "summary": "Chave invalida",
                "value": {
                  "error": "Invalid API key"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Recurso nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not found"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito (slug duplicado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "A post with this slug already exists"
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/admin/content/blog/posts/{id}/publish": {
      "post": {
        "tags": [
          "Blog Posts"
        ],
        "summary": "Publicar post",
        "description": "Altera o status do post para 'published' e define publishedAt como a data/hora atual. Retorna erro se ja estiver publicado.",
        "operationId": "publishBlogPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5abc123def456"
          }
        ],
        "responses": {
          "200": {
            "description": "Post publicado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                },
                "example": {
                  "id": "cm5abc123def456",
                  "title": "Como usar webhooks na Zé da API",
                  "slug": "como-usar-webhooks-na-ze-da-api",
                  "status": "published",
                  "publishedAt": "2026-02-21T14:30:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Post ja esta publicado",
            "content": {
              "application/json": {
                "example": {
                  "error": "Post is already published"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/admin/content/blog/posts/{id}/archive": {
      "post": {
        "tags": [
          "Blog Posts"
        ],
        "summary": "Arquivar post",
        "description": "Altera o status do post para 'archived'. Retorna erro se ja estiver arquivado.",
        "operationId": "archiveBlogPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5abc123def456"
          }
        ],
        "responses": {
          "200": {
            "description": "Post arquivado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                },
                "example": {
                  "id": "cm5abc123def456",
                  "status": "archived"
                }
              }
            }
          },
          "400": {
            "description": "Post ja esta arquivado",
            "content": {
              "application/json": {
                "example": {
                  "error": "Post is already archived"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/admin/content/blog/categories": {
      "get": {
        "tags": [
          "Blog Categories"
        ],
        "summary": "Listar categorias do blog",
        "description": "Retorna todas as categorias ordenadas por sortOrder, com contagem de posts.",
        "operationId": "listBlogCategories",
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogCategory"
                  }
                },
                "example": [
                  {
                    "id": "cm5cat001",
                    "name": "Tutoriais",
                    "slug": "tutoriais",
                    "description": "Guias passo a passo",
                    "sortOrder": 0,
                    "_count": {
                      "posts": 12
                    }
                  },
                  {
                    "id": "cm5cat002",
                    "name": "Novidades",
                    "slug": "novidades",
                    "description": "Lancamentos e updates",
                    "sortOrder": 1,
                    "_count": {
                      "posts": 8
                    }
                  },
                  {
                    "id": "cm5cat003",
                    "name": "Casos de Uso",
                    "slug": "casos-de-uso",
                    "description": null,
                    "sortOrder": 2,
                    "_count": {
                      "posts": 5
                    }
                  }
                ]
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Blog Categories"
        ],
        "summary": "Criar categoria do blog",
        "description": "Cria uma nova categoria. O slug e gerado automaticamente a partir do nome.",
        "operationId": "createBlogCategory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Integracoes"
                  },
                  "description": {
                    "type": "string",
                    "example": "Artigos sobre integracoes com sistemas externos"
                  },
                  "sortOrder": {
                    "type": "integer",
                    "example": 3
                  }
                }
              },
              "example": {
                "name": "Integracoes",
                "description": "Artigos sobre integracoes com sistemas externos",
                "sortOrder": 3
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogCategory"
                },
                "example": {
                  "id": "cm5cat004",
                  "name": "Integracoes",
                  "slug": "integracoes",
                  "description": "Artigos sobre integracoes com sistemas externos",
                  "sortOrder": 3,
                  "_count": {
                    "posts": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nome obrigatorio",
            "content": {
              "application/json": {
                "example": {
                  "error": "name is required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/admin/content/blog/categories/{id}": {
      "patch": {
        "tags": [
          "Blog Categories"
        ],
        "summary": "Atualizar categoria do blog",
        "description": "Atualiza parcialmente uma categoria. Se o nome mudar, o slug e recalculado.",
        "operationId": "updateBlogCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5cat001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "name": "Tutoriais e Guias",
                "sortOrder": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogCategory"
                },
                "example": {
                  "id": "cm5cat001",
                  "name": "Tutoriais e Guias",
                  "slug": "tutoriais-e-guias",
                  "description": "Guias passo a passo para usar a plataforma",
                  "sortOrder": 0,
                  "createdAt": "2026-01-10T08:00:00.000Z",
                  "updatedAt": "2026-02-21T16:00:00.000Z",
                  "_count": {
                    "posts": 12
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Blog Categories"
        ],
        "summary": "Excluir categoria do blog",
        "description": "Exclui uma categoria. Falha se houver posts vinculados — mova ou exclua os posts primeiro.",
        "operationId": "deleteBlogCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "cm5cat001"
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria excluida",
            "content": {
              "application/json": {
                "example": {
                  "success": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Categoria tem posts vinculados",
            "content": {
              "application/json": {
                "example": {
                  "error": "Cannot delete category with 12 post(s). Reassign or delete posts first."
                }
              }
            }
          }
        }
      }
    },
...
  }
}